完成删除文章和评论接口，完善翻译功能

+++ bb_server/forums/api/topic/views.py
@@ -11,7 +11,7 @@
 # Description:
 # **************************************************************************
 from flask import Markup, redirect, render_template, request, url_for, current_app
-from flask_babel import refresh, gettext as _
+from flask_babel import gettext as _
 from flask_login import current_user, login_required
 
 from flask_auth.form import form_validate
@@ -177,7 +177,7 @@ class TopicListView(MethodView):
         topic['author'] = user.username
         topic['is_good'] = 0
         topic['is_bad'] = 0
-        topic['diff_time'] = '0秒'
+        topic['diff_time'] = _('0 second')
         topic['bool_delete'] = bool_delete(user)
         #topic.board.topic_count = 1
         #topic.board.post_count = 1
@@ -204,12 +204,10 @@ class TopicView(MethodView):
         replies = []
         for i in reply: 
             user = i.author
-            #user = User.query.filter_by(id = i.author_id).first()
             diff_time = time_diff(i.updated_at)
             i.created_at = str(i.created_at)
             i.updated_at = str(i.updated_at)
             replies_data = object_as_dict(i)
-            #replies_data['content'] = json.loads(replies_data['content'])
             json_loads(replies_data, ['content', 'reference'])
             replies_data['author'] = user.username
             replies_data['diff_time'] = diff_time
@@ -220,7 +218,6 @@ class TopicView(MethodView):
             replies.append(replies_data)
         topic_user = topic.author   
         topic_data = object_as_dict(topic)
-        #topic_data['content'] = json.loads(topic_data['content'])
         json_loads(topic_data, ['content', 'title'])
         topic_data['author'] = topic_user.username
         topic_data['diff_time'] = diff_time
@@ -230,38 +227,24 @@ class TopicView(MethodView):
         topic_data['bool_delete'] = bool_delete(topic_user)
         Avatar(topic_data, topic_user)
         data = {
-            #'title': topic['title'],
-            #'form': object_as_dict(form),
             'topic': topic_data,
             'replies': replies,
             'replies_count': reply_count,
             'page_count': page_count
         }
-        #topic.read_count = 1
         msg = _('Topic Details')
         return get_json(1, msg, data)
         #return render_template('topic/topic.html', **data)
 
-    @form_validate(form_board)
-    def put(self, topicId):
-        form = form_board()
-        post_data = form.data
-        topic = Topic.query.filter_by(id=topicId).first_or_404()
-        title = post_data.pop('title', None)
-        content = post_data.pop('content', None)
-        content_type = post_data.pop('content_type', None)
-        category = post_data.pop('category', None)
-        if title is not None:
-            topic.title = title
-        if content is not None:
-            topic.content = content
-        if content_type is not None:
-            topic.content_type = content_type
-        if category is not None:
-            topic.board_id = int(category)
-        topic.save()
-        return HTTPResponse(HTTPResponse.NORMAL_STATUS).to_response()
-
+    def post(self, topicId):
+        user = request.user
+        topic = Topic.query.filter_by(id=topicId).first()
+        if user != topic.author:
+            msg = _('You do not have permission to delete this article.')
+            return get_json(0, msg, {})
+        topic.delete()
+        return get_json(1, _('success'), {})
+        
 
 class ReplyListView(MethodView):
     def get(self, topicId, page):
@@ -323,7 +306,7 @@ class ReplyListView(MethodView):
         replies_data['author'] = user.username
         replies_data['is_good'] = 0 
         replies_data['is_bad'] = 0
-        replies_data['diff_time'] = '0秒'
+        replies_data['diff_time'] = _('0 second')
         replies_data['replies_count'] = reply_count
         replies_data['bool_delete'] = bool_delete(user)
         # noticetopicId
@@ -334,77 +317,19 @@ class ReplyListView(MethodView):
         msg = _('success')
         return get_json(1, msg, replies_data)
 
-'''   
 class ReplyView(MethodView):
-    
     decorators = (reply_permission, )
+
     def post(self, replyId):
-        reply = Reply.query.filter_by(id=replyId).first_or_404()
-        reply_data = dict()
-        reply_data['author'] = User.query.filter_by(id = reply.author_id).first().username
-        reply_data['content'] = reply.content
-        post_data = request.data
         user = request.user
-        content = post_data.pop('content', None)
-        comment = Reply(content=content, topic_id = reply.id)
-        comment.author_id = user.id
-        comment.save()
-        diff_time = time_diff(comment.updated_at)
-        comment.created_at = str(comment.created_at)
-        comment.updated_at = str(comment.updated_at)
-        comment_data = object_as_dict(comment)
-        Avatar(comment_data, user)
-        comment_data['author'] = user.username
-        comment_data['is_good'] = 0
-        comment_data['is_bad'] = 0
-        data = {'reply_data':reply_data, 'comment_data':comment_data}
-        return get_json(1, '回复成功', data)
-
-
-class ReplyView(MethodView):
-
-    decorators = (reply_permission, )
-
-    def put(self, replyId):
-        post_data = request.data
-        reply = Reply.query.filter_by(id=replyId).first_or_404()
-        content = post_data.pop('content', None)
-        if content is not None:
-            reply.content = content
-        reply.save()
-        return HTTPResponse(HTTPResponse.NORMAL_STATUS).to_response()
-
-    def delete(self, replyId):
-        reply = Reply.query.filter_by(id=replyId).first_or_404()
+        reply = Reply.query.filter_by(id = replyId).first()
+        if user != reply.author:
+            msg = _('You do not have permission to delete this reply.')
+            return get_json(0, msg, {})
         reply.delete()
-        return HTTPResponse(HTTPResponse.NORMAL_STATUS).to_response()
-'''
-
-class LikeView(MethodView):
-
-    decorators = (like_permission, )
-
-    def post(self, replyId):
-        user = request.user
-        reply = Reply.query.filter_by(id=replyId).first_or_404()
-        reply.likers.append(user)
-        reply.save()
-        MessageClient.like(reply)
-        serializer = Serializer(reply, many=False)
-        return HTTPResponse(
-            HTTPResponse.NORMAL_STATUS, data=serializer.data).to_response()
-
-    def delete(self, replyId):
-        user = request.user
-        reply = Reply.query.filter_by(id=replyId).first_or_404()
-        reply.likers.remove(user)
-        reply.save()
-        serializer = Serializer(reply, many=False)
-        return HTTPResponse(
-            HTTPResponse.NORMAL_STATUS, data=serializer.data).to_response()
+        return get_json(1, _('success'), {})
 
 class ThumbView(MethodView):
-    #decorators = (thumd_permission, )
 
     def get(self, id, thumb):
         user = request.user
